sdm660-common: Enable adaptive suspend
This CL adds build properties enabling Adaptive Suspend's short
suspend backoff with a short suspend threshold of 2 seconds and
a max backoff time of 40s.  This is the same configuration that
was shipped (and remains active) on gs101-based devices.

Experimentation with these settings is being tracked on the bug,
and this CL is being prepared now so we can start to soak in
dogfood and observe the real-world behavior.

Bug: 284766768
Test: booted with a test kernel and used alarmtimer mods to
exercise high-frequency wakeup behavior

Change-Id: Ia0fd3b08058390ea9e70926455c7c6d460320d8a
Signed-off-by: GhostMaster69-dev <rathore6375@gmail.com>